requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
idna==3.4
lxml==4.9.3
prometheus-client==0.17.1
pyaml==23.7.0
pycparser==2.21
pyOpenSSL==23.2.0
python-consul2==0.1.5
pythonping==1.1.4
pyvmomi==8.0.1.0.2
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
sub==0.0.9
urllib3==2.0.4
